refactor(web-provider): fold the platform error mapper into its only consumer
mapPlatformWebError had one caller and its own test file. Exporting it from
platform-web-provider.ts keeps it a pure, directly-tested function while dropping
two files. provider-id-schema.ts stays in its own file, as the repo asks of schemas.
